Refacto App component because is too big
