+++ src/layouts/DockableLayout.jsx
@@ -1,0 +1,8 @@
+/**
+ * Dockable layout wrapper
+ */
+import DockableApp from '../DockableApp.jsx';
+
+export default function DockableLayout(props) {
+  return <DockableApp {...props} />;
+}
